refactor(api): Reorganization of dto/ride package
