+++ backend/FleetForge/src/main/java/com/team18/FleetForge/dto/ride/estimate/RideEstimateResponseDTO.java
@@ -1,4 +1,4 @@
-package com.team18.FleetForge.dto.ride;
+package com.team18.FleetForge.dto.ride.estimate;
 
 import com.team18.FleetForge.dto.RouteDTO;
 import lombok.Data;
