+++ backend/FleetForge/src/main/java/com/team18/FleetForge/dto/ride/lifecycle/RideCancellationRequestDTO.java
@@ -1,4 +1,4 @@
-package com.team18.FleetForge.dto.ride;
+package com.team18.FleetForge.dto.ride.lifecycle;
 
 import com.team18.FleetForge.model.enums.RideCancellationRole;
 import lombok.Data;
