+++ backend/FleetForge/src/main/java/com/team18/FleetForge/dto/ride/view/AdminRideHistoryItemDTO.java
@@ -1,4 +1,4 @@
-package com.team18.FleetForge.dto.ride;
+package com.team18.FleetForge.dto.ride.view;
 
 import com.team18.FleetForge.dto.UserSummaryDTO;
 import lombok.Builder;
